fix #15
The bug occurs, because PriorityQueue.put() already executes a comparison of the tupels in the queue. We don't really need to prioritize any snapshot, if two snapshots have exactly the same time, because they will most likely be equal. The do_saving() method of the SavingAgentController will discard snapshots with an older or equal time stamp (compared to the timestamp of the last saved data snapshot) anyway. So to allow the queue some sorting, even though it's not important for us, we added a uuid to the queue. This will ensure, that one of the snapshots will always be evaluated as "smaller", and therefore .put() and .get() will work.