implement alfred.run()
The function alfred.run() can be used to run the script.py directly, without the need of a run.py.

It works both when generate_experiment is defined as a class method and when it is defined as a standalone function.

Using it only requires to close the script.py with one of the following, depending on the script.py style:

run(generate_experiment)
run(script.generate_experiment)