implement sessionid
Previously, there was only and Experiment._uuid, which really was a session id, because it is a different id for each experiment startup.

For use in Mortimer, it is convenient though to have both: A unique id for an experiment, and a unique id for a session.

Thus, the Experiment._uuid is now an experiment identifiert when used with mortimer, and the Experiment._sessionid is (always) a session identifier. When alfred is used locally, the Experiment._uuid does not function as a unique identifier, in this case it still is a session id. Mortimer uses the Experiment.update() method to set the Experiment._uuid.