implement parameter loopvals
All Page classes now have the parameter "values", which is supposed to be a dictionary.

This parameter can be used to pass values to a page on intialization, which is useful if a page class is defined outside of generate_experiment() (the future default way). Through the parameter "values", values can be passed on to a page in a loop and accessed via self.values["key"].