custom settings as an init parameter for alfred.Experiment
This eliminates the need for alfred.Experiment.update(), it is altogether a more clean option.

Combined with this change is the introduction of an experiment ID for offline experiments aswell.  This ID is defined in config.conf as exp_id and can be used to bring together datasets that belong to the same experiment.