update local saving agent
The local saving agent will now check if the path specified in config.conf is an absolute path. If not, it will treat it as a relative path from inside the experiment directory. Previously it used the current working directory. This will fix unexpected behavior, when the working directory does not equal the experiment directory.